basic/ops.py
from typing import List, Tuple, Dict, Set

def get_list_len(l: List):
    l_out = None
    # 请在下一行编写代码，获取列表l的长度

    return l_out

def get_list_append_1(l: List):
    # 请在下一行编写代码，在l后面增加一个元素1

    return l

def get_list_reverse(l: List):
    # 请在下一行编写代码，将l的列表元素逆序

    return l

def get_list_exists(l:List, a):
    ret = None
    # 请在下一行编写代码，将判断a是否存在于l

    return ret

def judge_equal(a, b):
    ret = None
    # 请在下一行编写代码，将判断a b是否相等的结果反馈

    return ret

def return_large(a, b):
    ret = None
    # 请在下一行编写代码，返回较大的一个值；若两个类型不同，则返回None

    return ret

def list_select_int(l:list, a):
    l_out = None
    # 请在下一行编写代码，将l中所有的a取出，用取出的a形成新的列表并返回

    return l_out

def get_dict_keys(d:Dict):
    out = None
    # 请在下一行编写代码，将d中的所有key取出，并形成set后返回

    return out

def get_dict_values(d:Dict):
    out = None
    # 请在下一行编写代码，将d中的所有value取出，并形成set后返回

    return out

def judge_dict_key_exists(d:Dict, k):
    ret = None
    # 请在下一行编写代码，判断k是否在d的key中,如果存在返回True,否则返回False

    return ret

def judge_dict_value_exists(d:Dict, v):
    ret = None
    # 请在下一行编写代码，判断k是否在d的value中,如果存在返回True,否则返回False

    return ret

def remove_dict_key(d:Dict, k):
    ret = None
    # 请在下一行编写代码，判断k是否在d的key中,如果存在则删除key

    return ret

def update_dict(d:Dict, d1:Dict):
    ret = None
    # 请在下一行编写代码，使用d1所有内容更新d,并将最后结果返回

    return ret

def dict_get(d:Dict, k):
    ret = None
    # 请在下一行编写代码，使用k作为key，获取d中的value；如果存在则返回value，如果不存在返回字符串 "abc"

    return ret

def set_add(a:Set, b:Set):
    ret = None
    # 请在下一行编写代码，将set b中所有元素加到set a当中，并返回a

    return ret

def set_remove(a:Set, b:Set):
    ret = None
    # 请在下一行编写代码，遍历set b中所有元素，若存在于set a中则删除该元素；最后返回a

    return ret

def list_add_sub(a:List):
    ret = None
    # 请在下一行编写代码，遍历list a, 若元素大于0，则将其+1；若元素小于0增将其-1;若元素等于0，则将其变为255；若无元素，则返回None

    return ret

def list_plus_2(a:int, b:int):
    ret = None
    # 请在下一行编写代码，返回结果ret是一个list; 输出从a开始（包括a）自增，并将自增后的值加入到list中。每次自增2，直到自增数大于b停止

    return ret


if __name__ == "__main__":
    ret = get_list_len()